Split SagaInstancesRegistry into a read-only interface and an internal impl
The registry was a public class in a public package with a public register method,
kept honest only by a javadoc line asking applications not to call it. That is a
footgun documented rather than prevented, on a bean applications are told to inject,
and it contradicted SagaInstances in the same change advertising no way to write.

Unlike SagaStateStore.compareAndSave, which an application legitimately calls when
wiring its own store, register has no legitimate application caller, so the write
side is not public API. SagaInstancesRegistry is now a read-only interface with
find/get/sagaIds; SagaInstancesRegistryImpl carries register and lives in
org.occurrent.dsl.saga.internal, whose package name is what signals not-user-API.
This is the same split ADR 63's FlowState uses, and FlowStateImpl only gets away
with a public mutable surface because it sits in an internal package too.

The starter constructs the impl, declares the bean as the interface, and resolves the
concrete type when it needs to write, so a user-replaced registry bean stays empty and
warns instead of failing a saga. Behaviour is otherwise unchanged: same duplicate-id
rejection, same id-listing message, same refresh-timing documentation, and registry
population stays independent of the ConfigurableApplicationContext singleton path.

Also gives org.occurrent.dsl.saga.internal the package-info javadoc it was missing,
matching flow.internal.
